workers/media/requirements.txt
# YouTube Production Assistant - Media Worker Dependencies

# Database
psycopg2-binary==2.9.9
supabase==2.9.1

# Video Processing (using ffmpeg directly via subprocess - moviepy not needed)
ffmpeg-python==0.2.0

# Audio Analysis
webrtcvad==2.0.10
pydub==0.25.1
librosa==0.10.1

# Silero VAD (Neural Voice Activity Detection)
torch==2.1.2
torchaudio==2.1.2

# Transcription (using openai-whisper only - simpler, no native deps)
openai-whisper==20231117

# LLM for retake analysis
openai==1.12.0

# Face/Pose Analysis
mediapipe==0.10.9
opencv-python==4.9.0.80
numpy==1.26.3
Pillow==10.2.0

# HTTP
httpx==0.27.0

# Utilities
python-dotenv==1.0.0

